+++ c/frame/py_frame.cc
@@ -19,6 +19,7 @@ namespace py {
 
 PKArgs Frame::Type::args___init__(1, 0, 3, false, false,
                                   {"src", "names", "stypes", "stype"});
+PKArgs Frame::Type::args_colindex(1, 0, 0, false, false, {"name"});
 
 
 const char* Frame::Type::classname() {
@@ -37,10 +38,12 @@ const char* Frame::Type::classdoc() {
     "This is a primary data structure for the `datatable` module.\n";
 }
 
+
 void Frame::Type::init_getsetters(GetSetters& gs)
 {
   gs.add<&Frame::get_ncols>("ncols",
     "Number of columns in the Frame\n");
+
   gs.add<&Frame::get_nrows, &Frame::set_nrows>("nrows",
     "Number of rows in the Frame.\n"
     "\n"
@@ -49,24 +52,72 @@ void Frame::Type::init_getsetters(GetSetters& gs)
     "current, or filling with NAs if the new number of rows is greater.\n"
     "\n"
     "Increasing the number of rows of a keyed Frame is not allowed.\n");
+
   gs.add<&Frame::get_shape>("shape",
     "Tuple with (nrows, ncols) dimensions of the Frame\n");
+
+  gs.add<&Frame::get_stypes>("stypes",
+    "The tuple of each column's stypes (\"storage types\")\n");
+
+  gs.add<&Frame::get_ltypes>("ltypes",
+    "The tuple of each column's ltypes (\"logical types\")\n");
+
+  gs.add<&Frame::get_names, &Frame::set_names>("names",
+    "Tuple of column names.\n"
+    "\n"
+    "You can rename the Frame's columns by assigning a new list/tuple of\n"
+    "names to this property. The length of the new list of names must be\n"
+    "the same as the number of columns in the Frame.\n"
+    "\n"
+    "It is also possible to rename just a few columns by assigning a\n"
+    "dictionary ``{oldname: newname, ...}``. Any column not listed in the\n"
+    "dictionary will retain its name.\n"
+    "\n"
+    "Examples\n"
+    "--------\n"
+    ">>> d0 = dt.Frame([[1], [2], [3]])\n"
+    ">>> d0.names = ['A', 'B', 'C']\n"
+    ">>> d0.names\n"
+    "('A', 'B', 'C')\n"
+    ">>> d0.names = {'B': 'middle'}\n"
+    ">>> d0.names\n"
+    "('A', 'middle', 'C')\n"
+    ">>> del d0.names\n"
+    ">>> d0.names\n"
+    "('C0', 'C1', 'C2')");
+
   gs.add<&Frame::get_key>("key");
   gs.add<&Frame::get_internal>("internal", "[DEPRECATED]");
   gs.add<&Frame::get_internal, &Frame::set_internal>("_dt");
 }
 
-void Frame::Type::init_methods(Methods&) {
-  // mm.add<&Frame::bang, args_bang>("bang");
+
+void Frame::Type::init_methods(Methods& mm) {
+  mm.add<&Frame::colindex, args_colindex>("colindex",
+    "colindex(self, name)\n"
+    "--\n\n"
+    "Return index of the column ``name``.\n"
+    "\n"
+    ":param name: name of the column to find the index for. This can also\n"
+    "    be an index of a column, in which case the index is checked that\n"
+    "    it doesn't go out-of-bounds, and negative index is converted into\n"
+    "    positive.\n"
+    ":raises ValueError: if the requested column does not exist.\n");
 }
 
 
 
+//------------------------------------------------------------------------------
+// Misc
+//------------------------------------------------------------------------------
+
 void Frame::m__dealloc__() {
   Py_XDECREF(core_dt);
-  // `dt` is already managed by `core_dt`.
-  // delete dt;
-  dt = nullptr;
+  Py_XDECREF(stypes);
+  Py_XDECREF(ltypes);
+  Py_XDECREF(names);
+  Py_XDECREF(inames);
+  dt = nullptr;  // `dt` is already managed by `core_dt`
 }
 
 void Frame::m__get_buffer__(Py_buffer* , int ) const {
@@ -74,6 +125,7 @@ void Frame::m__get_buffer__(Py_buffer* , int ) const {
 
 void Frame::m__release_buffer__(Py_buffer*) const {
 }
+
 
 
 //------------------------------------------------------------------------------
@@ -109,6 +161,33 @@ oobj Frame::get_shape() const {
   return shape;
 }
 
+
+oobj Frame::get_stypes() const {
+  if (stypes == nullptr) {
+    py::otuple ostypes(dt->ncols);
+    for (size_t i = 0; i < ostypes.size(); ++i) {
+      SType st = dt->columns[i]->stype();
+      ostypes.set(i, info(st).py_stype());
+    }
+    stypes = ostypes.release();
+  }
+  return oobj(stypes);
+}
+
+
+oobj Frame::get_ltypes() const {
+  if (ltypes == nullptr) {
+    py::otuple oltypes(dt->ncols);
+    for (size_t i = 0; i < oltypes.size(); ++i) {
+      SType st = dt->columns[i]->stype();
+      oltypes.set(i, info(st).py_ltype());
+    }
+    ltypes = oltypes.release();
+  }
+  return oobj(ltypes);
+}
+
+
 oobj Frame::get_key() const {
   py::otuple key(dt->nkeys);
   // Fill in the keys...
@@ -120,9 +199,13 @@ oobj Frame::get_internal() const {
 }
 
 void Frame::set_internal(obj _dt) {
+  oobj tmp(_dt);  // In case the new _dt is same as, or related to the current
+                  // Frame, we do no want it to be deleted. By creating a new
+                  // reference to `_dt` we ensure it can't get deleted.
   m__dealloc__();
   dt = _dt.to_frame();
   core_dt = static_cast<pydatatable::obj*>(_dt.to_pyobject_newref());
+  core_dt->_frame = this;
 }