+++ c/frame/py_frame.h
@@ -11,17 +11,27 @@
 #include "datatable.h"
 #include "py_datatable.h"
 
+namespace pydatatable {  // temp
+  void _clear_types(obj*);
+  PyObject* check(obj*, PyObject*);
+}
+
+
 namespace py {
 
 
 /**
- * [WIP]
- * This class is exposed to python via `datatable.lib.core.Frame`, but it is
- * not directly usable yet.
+ * This class currently serves as a base for datatable.Frame, but eventually
+ * all functionality will be moved here, and this class will be the main
+ * user-facing Frame class.
  */
 class Frame : public PyObject {
   private:
     DataTable* dt;
+    mutable PyObject* stypes;  // memoized tuple of stypes
+    mutable PyObject* ltypes;  // memoized tuple of ltypes
+    mutable PyObject* names;   // memoized tuple of column names
+    mutable PyObject* inames;  // memoized dict of {column name: index}
 
     pydatatable::obj* core_dt;  // TODO: remove
 
@@ -29,7 +39,7 @@ class Frame : public PyObject {
     class Type : public ExtType<Frame> {
       public:
         static PKArgs args___init__;
-        static PKArgs args_test;
+        static PKArgs args_colindex;
         static const char* classname();
         static const char* classdoc();
         static bool is_subclassable() { return true; }
@@ -45,12 +55,28 @@ class Frame : public PyObject {
 
     oobj get_ncols() const;
     oobj get_nrows() const;
-    void set_nrows(obj);
     oobj get_shape() const;
+    oobj get_stypes() const;
+    oobj get_ltypes() const;
+    oobj get_names() const;
     oobj get_key() const;
     oobj get_internal() const;
-    void set_internal(obj _dt);
+    void set_nrows(obj);
+    void set_names(obj);
+    void set_internal(obj);
 
+    oobj colindex(PKArgs&);
+
+  private:
+    void _clear_names();
+    void _init_names() const;
+    void _init_inames() const;
+    void _fill_default_names();
+    void _dedup_and_save_names(py::list);
+    void _replace_names_from_map(py::odict);
+
+    friend void pydatatable::_clear_types(pydatatable::obj*); // temp
+    friend PyObject* pydatatable::check(pydatatable::obj*, PyObject*); // temp
 };