+++ c/python/tuple.h
@@ -13,6 +13,14 @@
 namespace py {
 
 
+/**
+ * Class representing a pythonic tuple of elements.
+ *
+ * This can be either a tuple passed down from Python, or a new tuple object.
+ * If `otuple` is created from an existing PyObject, its elements must not be
+ * modified. On the contrary, if a tuple is created from scratch, its elements
+ * must be properly initialized before the tuple can be returned to Python.
+ */
 class otuple : public oobj {
   public:
     using oobj::oobj;
@@ -21,6 +29,7 @@ class otuple : public oobj {
     otuple(int64_t n);
 
     size_t size() const;
+    obj  get(size_t i) const;
     void set(size_t i, const _obj& value);
     void set(size_t i, oobj&& value);
 };