+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   use a list or tuple of strings instead.
 - `Frame.resize()` was removed -- same functionality is available via
   assigning to `Frame.nrows`.
+- `Frame.rename()` was removed -- .name setter can be used instead.
 
 #### Fixed
 - bug in dt.cbind() where the first Frame in the list was ignored.